Extract direct association manipulation from VM. Mark `read_association` as internal.
